feat(research): added product and prices in app config
