+++ .env.template
@@ -21,3 +21,8 @@ NEXT_PUBLIC_API_DOMAIN=op://k33-web-env/$APP_ENV/domain/API
 #algolia
 NEXT_PUBLIC_ALGOLIA_APP_ID=op://k33-web-env/$APP_ENV/algolia/APP_ID
 NEXT_PUBLIC_ALGOLIA_SEARCH_API_KEY=op://k33-web-env/$APP_ENV/algolia/SEARCH_API_KEY
+
+# stripe products and prices
+NEXT_PUBLIC_STRIPE_PRODUCT_ID=
+NEXT_PUBLIC_STRIPE_MONTHLY_PRICE_ID=
+NEXT_PUBLIC_STRIPE_YEARLY_PRICE_ID=
