+++ apps/research/src/components/platform/index.ts
@@ -7,3 +7,4 @@ export { default as NamedDivider } from './NamedDivider';
 export { default as SectionHeader } from './SectionHeader';
 export * from './SectionHeader';
 export { default as SectionDescriptionHeader } from './SectionDescriptionHeader';
+export { default as PricingTable } from './PricingTable';
