+++ turbo.json
@@ -47,6 +47,9 @@
     "NEXT_PUBLIC_ALGOLIA_APP_ID",
 
     "NEXT_PUBLIC_WEB_DOMAIN",
+    "NEXT_PUBLIC_STRIPE_PRODUCT_ID",
+    "NEXT_PUBLIC_STRIPE_MONTHLY_PRICE_ID",
+    "NEXT_PUBLIC_STRIPE_YEARLY_PRICE_ID",
     "NEXT_PUBLIC_API_DOMAIN",
     "npm_lifecycle_event"
   ]
